AddressBook.java: Add sort command
No sort command exist. There is only the list command to list based on the index they are added in.

Since no sort command exist, user is not able to sort the list in alphabetical order.

Let's add a sort command such that user is able to print the list in alphabetical order.